# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/products/index": {
            "get": {
                "tags": [
                    "Product Management"
                ],
                "summary": "Get list of products",
                "operationId": "5946b89f5ead55364219e1cd090e6b7f",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search product by name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "food"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for paginated results",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/detail": {
            "get": {
                "tags": [
                    "Product Management"
                ],
                "summary": "Get detail of Product",
                "operationId": "9e63289bc62e240a202b71724cc796fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/create": {
            "post": {
                "tags": [
                    "Product Management"
                ],
                "summary": "Create a new Product",
                "operationId": "944a15ab2c7c2ff4f72d6e3e3296f626",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Iphone 16 promax"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Giao hàng rẻ nhưng không nhanh"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/edit": {
            "post": {
                "tags": [
                    "Product Management"
                ],
                "summary": "Update new Product",
                "operationId": "0d08c94ff8e082997ed88c2e46f25f81",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Hoa Loa kèn"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Hoa Loa kèn rất loa kèn"
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/delete": {
            "delete": {
                "tags": [
                    "Product Management"
                ],
                "summary": "Delete a Product",
                "operationId": "36773dd30c318a6b4036e11836570d9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Product ID is deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/detail-child": {
            "get": {
                "tags": [
                    "Product Management"
                ],
                "summary": "Get detail of Feature Product",
                "operationId": "14a98a2f1b0041f5aac8412dcef2f332",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id of Product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/create-child": {
            "post": {
                "tags": [
                    "Product Management"
                ],
                "summary": "Create a new Feature Product",
                "operationId": "bfc4e70940fc25701f36aa559710cc56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "feature_name": {
                                        "type": "string",
                                        "example": "Iphone 16 promax"
                                    },
                                    "cost_price": {
                                        "type": "integer",
                                        "example": 20000
                                    },
                                    "selling_price": {
                                        "type": "integer",
                                        "example": 30000
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/edit-child": {
            "post": {
                "tags": [
                    "Product Management"
                ],
                "summary": "Edit a Feature Product",
                "operationId": "162c68510e2d42eb6e94721d287b59cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "feature_name": {
                                        "type": "string",
                                        "example": "Iphone 16 promax"
                                    },
                                    "cost_price": {
                                        "type": "integer",
                                        "example": 20000
                                    },
                                    "selling_price": {
                                        "type": "integer",
                                        "example": 30000
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 200
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/products/delete-child": {
            "delete": {
                "tags": [
                    "Product Management"
                ],
                "summary": "Delete a Feature Product",
                "operationId": "b4764574ec6b80e26433b9f98b845534",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Product ID is deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}